backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
chromadb>=0.6.0
httpx>=0.28.0
python-dotenv>=1.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
